backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
werkzeug==2.2.3
torch>=1.9.0
transformers>=4.11.3
pillow>=8.3.1
opencv-python>=4.5.3
numpy>=1.21.2
timm>=1.0.0
